Created page for apply to vacancy

+++ src/actions/applicationActions.ts
@@ -2,11 +2,37 @@
 
 import { dbConnect } from "@/config/db";
 import ApplicationModel, { IApplication } from "@/models/Application";
+import { authOptions } from "@/app/api/auth/[...nextauth]/route";
+import { getServerSession } from "next-auth";
+import { IApplicationCreate } from "@/types/application";
 
-export async function createApplication(data: IApplication) {
-  await dbConnect();
-  const newApplication = new ApplicationModel(data);
-  return await newApplication.save();
+export async function createApplication(data: IApplicationCreate) {
+  try {
+    await dbConnect();
+
+    const session = await getServerSession(authOptions);
+
+    if (!session || !session.user) {
+      throw new Error("User is not authenticated");
+    }
+
+    const newApplication = new ApplicationModel({
+      ...data,
+      applicant: session.user.id,
+    });
+
+    await newApplication.save();
+
+    return {
+      success: true,
+      message: "Application submitted successfully",
+    };
+  } catch (error) {
+    return {
+      success: false,
+      error: error instanceof Error ? error.message : "Unknown error",
+    };
+  }
 }
 
 export async function getApplicationById(applicationId: string) {
@@ -16,9 +42,21 @@ export async function getApplicationById(applicationId: string) {
 
 export async function updateApplication(
   applicationId: string,
-  data: Partial<IApplication>
+  data: Partial<Omit<IApplication, "applicant">>
 ) {
   await dbConnect();
+
+  const session = await getServerSession(authOptions);
+
+  if (!session || !session.user) {
+    throw new Error("User is not authenticated.");
+  }
+
+  const application = await ApplicationModel.findById(applicationId);
+  if (application?.applicant.toString() !== session.user.id) {
+    throw new Error("You are not authorized to update this application.");
+  }
+
   return await ApplicationModel.findByIdAndUpdate(applicationId, data, {
     new: true,
   });
@@ -26,5 +64,17 @@ export async function updateApplication(
 
 export async function deleteApplication(applicationId: string) {
   await dbConnect();
+
+  const session = await getServerSession(authOptions);
+
+  if (!session || !session.user) {
+    throw new Error("User is not authenticated.");
+  }
+
+  const application = await ApplicationModel.findById(applicationId);
+  if (application?.applicant.toString() !== session.user.id) {
+    throw new Error("You are not authorized to delete this application.");
+  }
+
   return await ApplicationModel.findByIdAndDelete(applicationId);
 }
